+++ db/migrate/20221212232052_add_monitoring_cache_to_services_and_datasources.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class AddMonitoringCacheToServicesAndDatasources < ActiveRecord::Migration[6.1]
+  def change
+    add_column :services, :availability_cache, :float
+    add_column :services, :reliability_cache, :float
+    add_column :datasources, :availability_cache, :float
+    add_column :datasources, :reliability_cache, :float
+  end
+end