+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ## [3.45.0-milestone]
 
+### Added
+- EOSC Portal monitoring data to the service details view (@goreck888)
+
 ### Changed
 - AAI authorization endpoints (@goreck888)